WeatherApi.Infrastructure/Services/WeatherService.cs
using WeatherApi.Domain.Interfaces;
using WeatherApi.Domain.Models;

namespace WeatherApi.Infrastructure.Services;

public class WeatherService : IWeatherService
{
    private readonly string[] _summaries =
    [
        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
    ];

    public async Task<WeatherForecast[]> GetWeather(string location)
    {
        var result = Enumerable.Range(1, 5).Select(index =>
                new WeatherForecast
                (
                    DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
                    Random.Shared.Next(-20, 55),
                    _summaries[Random.Shared.Next(_summaries.Length)]
                ))
            .ToArray();

        return result;
    }
}